Levenberg-Marquardt (LM) complex fix
problem with LM for complex turned out to be problem with Jacobian * error.  Used the gradient way of calculating Jacobian * error, since this propagates the error real,im parts correctly.  For the Hessian = Jacobian * Jacobian', the LM computation was correct already since this involves the propagation of the function and current weights derivatives and does not involve the current error.
Also updated some examples