fixed sign error in deltax,deltaf, qr iteration
look for -1* which needs to be there to subtract the two terms that go with imag(f) * 
added extra step to squeeze out residual, maybe not worth it

removed the parfor, and used matrix ops to do the backprop.  Well, it's ok for single outputs.  When dealing with multiple outputs, will have to change things